requirements.txt
requests
beautifulsoup4
torch
torchvision
transformers
pillow
python-dotenv